Move prewarmStyleIcons into styleParser module
